+++ activerecord/test/migrations/valid_do_nothing/2_valid_do_nothing_second.rb
@@ -1,0 +1,7 @@
+class ValidDoNothingSecond < ActiveRecord::Migration
+  def self.up
+  end
+
+  def self.down
+  end
+end